requirements.txt
google-genai>=0.8.0
opencv-python>=4.9.0
requests>=2.31.0
pyyaml>=6.0.1
numpy>=1.26.0
flask>=3.0.0
pytest>=8.0.0
